move commands into moc_sprint_tools.cmd module
move the following commands into the moc_sprint_tools.cmd module:

- cards-missing-from-backlog
- close-sprint-boards
- create-sprint-boards
- label-cards-in-sprint
- label-needs-description
- sort-cards-by-priority

The commands in utils.py stay in the top level module.

@feature/tests
