+++ moc_sprint_tools/cli.py
@@ -4,13 +4,14 @@ import logging
 import os
 
 from moc_sprint_tools import defaults
-from moc_sprint_tools import label_cards_in_sprint
-from moc_sprint_tools import label_needs_description
-from moc_sprint_tools import sort_cards_by_priority
-from moc_sprint_tools import cards_missing_from_backlog
-from moc_sprint_tools import close_sprint_boards
-from moc_sprint_tools import create_sprint_boards
 from moc_sprint_tools import utils
+
+from moc_sprint_tools.cmd import cards_missing_from_backlog
+from moc_sprint_tools.cmd import close_sprint_boards
+from moc_sprint_tools.cmd import create_sprint_boards
+from moc_sprint_tools.cmd import label_cards_in_sprint
+from moc_sprint_tools.cmd import label_needs_description
+from moc_sprint_tools.cmd import sort_cards_by_priority
 
 from moc_sprint_tools.sprintman import Sprintman
